Check if the element is exist in MouseScrollTo operation and if it's not - just skip the operation.
The main problem is the isDisplayed() method is not present in the W3C WebDriver standard. Therefore it behaves differently in different browsers. In Firefox an element is considered hidden if it's not displayed in the browser viewport. It's a really anal behavior. To mitigate it in the Visible condition we check if the element exists in the page source, then we scroll to the element before checking if it's displaying. But if the element disappears in a short time window between Exist and MouseScrollTo operations than the running test could fail with an exception. So we again check if the element exists in the MouseScrollTo operation to avoid this failure. It's not a 100% working method, but it should work in most cases.